Convert all Swagger GitHub release tests to OpenAPI
