+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/GitHubWorkflowIT.java
@@ -46,10 +46,8 @@ import io.dropwizard.client.JerseyClientBuilder;
 import io.openapi.model.DescriptorType;
 import io.swagger.client.ApiClient;
 import io.swagger.client.ApiException;
-import io.swagger.client.api.UsersApi;
 import io.swagger.client.api.WorkflowsApi;
 import io.swagger.client.model.Image;
-import io.swagger.client.model.PublishRequest;
 import io.swagger.client.model.Workflow;
 import io.swagger.client.model.WorkflowVersion;
 import jakarta.ws.rs.client.Client;
@@ -114,16 +112,15 @@ class GitHubWorkflowIT extends BaseIT {
      */
     @Test
     void testPublishingAndListingOfPublished() throws ApiException {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
-        io.dockstore.openapi.client.ApiClient openAPIWebClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        final io.dockstore.openapi.client.ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        io.dockstore.openapi.client.api.WorkflowsApi workflowApi = new io.dockstore.openapi.client.api.WorkflowsApi(webClient);
 
         // should start with nothing published
         assertTrue(workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null).isEmpty(), "should start with nothing published ");
         // refresh just for the current user
-        UsersApi usersApi = new UsersApi(webClient);
+        io.dockstore.openapi.client.api.UsersApi usersApi = new io.dockstore.openapi.client.api.UsersApi(webClient);
 
-        refreshByOrganizationReplacement(workflowApi, openAPIWebClient);
+        refreshByOrganizationReplacement(workflowApi, webClient);
 
         assertTrue(workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null).isEmpty(), "should remain with nothing published ");
         // ensure that sorting or filtering don't expose unpublished workflows
@@ -133,68 +130,68 @@ class GitHubWorkflowIT extends BaseIT {
 
         // assertTrue("should have a bunch of stub workflows: " +  usersApi..allWorkflows().size(), workflowApi.allWorkflows().size() == 4);
 
-        final Workflow workflowByPath = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, BIOWORKFLOW, null);
+        final io.dockstore.openapi.client.model.Workflow workflowByPath = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, WorkflowSubClass.BIOWORKFLOW, null);
         // refresh targeted
-        workflowApi.refresh(workflowByPath.getId(), false);
+        workflowApi.refresh1(workflowByPath.getId(), false);
 
         // publish one
-        final PublishRequest publishRequest = CommonTestUtilities.createPublishRequest(true);
-        workflowApi.publish(workflowByPath.getId(), publishRequest);
+        final io.dockstore.openapi.client.model.PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        workflowApi.publish1(workflowByPath.getId(), publishRequest);
         assertEquals(1, workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null).size(),
             "should have one published, found  " + workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null).size());
-        final Workflow publishedWorkflow = workflowApi.getPublishedWorkflow(workflowByPath.getId(), null);
+        final io.dockstore.openapi.client.model.Workflow publishedWorkflow = workflowApi.getPublishedWorkflow(workflowByPath.getId(), null);
         assertNotNull(publishedWorkflow, "did not get published workflow");
-        final Workflow publishedWorkflowByPath = workflowApi
-            .getPublishedWorkflowByPath(DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, BIOWORKFLOW, null,  null);
+        final io.dockstore.openapi.client.model.Workflow publishedWorkflowByPath = workflowApi
+            .getPublishedWorkflowByPath(DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, WorkflowSubClass.BIOWORKFLOW, null,  null);
         assertNotNull(publishedWorkflowByPath, "did not get published workflow");
 
         // publish everything so pagination testing makes more sense (going to unfortunately use rate limit)
         Lists.newArrayList("github.com/" + DOCKSTORE_TEST_USER_2_HELLO_DOCKSTORE_NAME,
                 "github.com/DockstoreTestUser2/dockstore-whalesay-imports", "github.com/DockstoreTestUser2/parameter_test_workflow")
             .forEach(path -> {
-                Workflow workflow = workflowApi.getWorkflowByPath(path, BIOWORKFLOW, null);
-                workflowApi.refresh(workflow.getId(), false);
-                workflowApi.publish(workflow.getId(), publishRequest);
+                io.dockstore.openapi.client.model.Workflow workflow = workflowApi.getWorkflowByPath(path, WorkflowSubClass.BIOWORKFLOW, null);
+                workflowApi.refresh1(workflow.getId(), false);
+                workflowApi.publish1(workflow.getId(), publishRequest);
             });
-        List<Workflow> workflows = workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null);
+        List<io.dockstore.openapi.client.model.Workflow> workflows = workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null);
         // test offset
         assertEquals(workflowApi.allPublishedWorkflows(1, null, null, null, null, false, null).get(0).getId(), workflows.get(1).getId(), "offset does not seem to be working");
         // test limit
         assertEquals(1, workflowApi.allPublishedWorkflows(null, 1, null, null, null, false, null).size());
         // test custom sort column
-        List<Workflow> ascId = workflowApi.allPublishedWorkflows(null, null, null, "id", "asc", false, null);
-        List<Workflow> descId = workflowApi.allPublishedWorkflows(null, null, null, "id", "desc", false, null);
+        List<io.dockstore.openapi.client.model.Workflow> ascId = workflowApi.allPublishedWorkflows(null, null, null, "id", "asc", false, null);
+        List<io.dockstore.openapi.client.model.Workflow> descId = workflowApi.allPublishedWorkflows(null, null, null, "id", "desc", false, null);
         assertEquals(ascId.get(0).getId(), descId.get(descId.size() - 1).getId(), "sort by id does not seem to be working");
         // test filter
-        List<Workflow> filteredLowercase = workflowApi.allPublishedWorkflows(null, null, "whale", "stars", null, false, null);
+        List<io.dockstore.openapi.client.model.Workflow> filteredLowercase = workflowApi.allPublishedWorkflows(null, null, "whale", "stars", null, false, null);
         assertEquals(1, filteredLowercase.size());
         filteredLowercase.forEach(workflow -> assertNull(workflow.getAliases()));
-        List<Workflow> filteredUppercase = workflowApi.allPublishedWorkflows(null, null, "WHALE", "stars", null, false, null);
+        List<io.dockstore.openapi.client.model.Workflow> filteredUppercase = workflowApi.allPublishedWorkflows(null, null, "WHALE", "stars", null, false, null);
         assertEquals(1, filteredUppercase.size());
         assertEquals(filteredLowercase, filteredUppercase);
 
         // Tests for subclass
 
         assertEquals(0, workflowApi.allPublishedWorkflows(null, null, null, null, null, false,
-            WorkflowSubClass.APPTOOL.getValue()).size(), "There should be no app tools published");
+            WorkflowSubClass.APPTOOL).size(), "There should be no app tools published");
 
         final int publishedWorkflowsCount = workflowApi.allPublishedWorkflows(null, null, null, null, null, false,
             null).size();
         assertEquals(publishedWorkflowsCount, workflowApi.allPublishedWorkflows(null, null, null, null, null, false,
-            WorkflowSubClass.BIOWORKFLOW.getValue()).size(), "An null subclass param defaults to services param value");
+            WorkflowSubClass.BIOWORKFLOW).size(), "An null subclass param defaults to services param value");
 
         // Create an app tool and publish it
         GitHubAppHelper.registerAppTool(webClient);
-        Workflow appTool = workflowApi.getWorkflowByPath("github.com/" + toolAndWorkflowRepoToolPath, APPTOOL, "versions");
-        workflowApi.publish(appTool.getId(), publishRequest);
+        io.dockstore.openapi.client.model.Workflow appTool = workflowApi.getWorkflowByPath("github.com/" + toolAndWorkflowRepoToolPath, WorkflowSubClass.APPTOOL, "versions");
+        workflowApi.publish1(appTool.getId(), publishRequest);
         assertEquals(1, workflowApi.allPublishedWorkflows(null, null, null, null, null, false,
-            WorkflowSubClass.APPTOOL.getValue()).size(), "There should be 1 app tool published");
+            WorkflowSubClass.APPTOOL).size(), "There should be 1 app tool published");
         // there should be one app tool account to header count too
         Client jerseyClient = new JerseyClientBuilder(SUPPORT.getEnvironment()).build("test client");
         CommonTestUtilities.testXTotalCount(jerseyClient, String.format("http://localhost:%d/workflows/published?subclass=APPTOOL", SUPPORT.getLocalPort()), 1);
 
         assertEquals(publishedWorkflowsCount, workflowApi.allPublishedWorkflows(null, null, null, null, null, false,
-            WorkflowSubClass.BIOWORKFLOW.getValue()).size(), "Published workflow count should be unchanged");
+            WorkflowSubClass.BIOWORKFLOW).size(), "Published workflow count should be unchanged");
     }
 
     /**
@@ -203,19 +200,19 @@ class GitHubWorkflowIT extends BaseIT {
      */
     @Test
     void testGetPublishedWorkflowsWithInvalidSortCol() {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
-        final PublishRequest publishRequest = CommonTestUtilities.createPublishRequest(true);
+        final io.dockstore.openapi.client.ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        io.dockstore.openapi.client.api.WorkflowsApi workflowApi = new io.dockstore.openapi.client.api.WorkflowsApi(webClient);
+        final io.dockstore.openapi.client.model.PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
 
         GitHubAppHelper.registerAppTool(webClient);
-        Workflow appTool = workflowApi.getWorkflowByPath("github.com/" + toolAndWorkflowRepoToolPath, APPTOOL, "versions");
-        workflowApi.publish(appTool.getId(), publishRequest);
+        io.dockstore.openapi.client.model.Workflow appTool = workflowApi.getWorkflowByPath("github.com/" + toolAndWorkflowRepoToolPath, WorkflowSubClass.APPTOOL, "versions");
+        workflowApi.publish1(appTool.getId(), publishRequest);
 
         assertEquals(1, workflowApi.allPublishedWorkflows(null, null, null, null, null, false,
-                WorkflowSubClass.APPTOOL.getValue()).size(), "There should be 1 app tool published");
+                WorkflowSubClass.APPTOOL).size(), "There should be 1 app tool published");
 
-        ApiException exception = assertThrows(ApiException.class, () -> workflowApi.allPublishedWorkflows(null, null, null, "invalid", null, false,
-                    WorkflowSubClass.APPTOOL.getValue()));
+        io.dockstore.openapi.client.ApiException exception = assertThrows(io.dockstore.openapi.client.ApiException.class, () -> workflowApi.allPublishedWorkflows(null, null, null, "invalid", null, false,
+                    WorkflowSubClass.APPTOOL));
         assertTrue(exception.getMessage().contains(INVALID_SORTCOL_MESSAGE));
         assertEquals(HttpStatus.SC_BAD_REQUEST, exception.getCode(), "There should be a 400 error");
     }
