+++ dockstore-integration-testing/src/test/java/io/dockstore/webservice/WebhookIT.java
@@ -1,10 +1,15 @@
 
 package io.dockstore.webservice;
 
+import static io.dockstore.client.cli.WorkflowIT.DOCKSTORE_TEST_USER_2_HELLO_DOCKSTORE_NAME;
 import static io.dockstore.common.Hoverfly.ORCID_SIMULATION_SOURCE;
+import static io.dockstore.webservice.Constants.DOCKSTORE_YML_PATH;
 import static io.dockstore.webservice.helpers.GitHubAppHelper.INSTALLATION_ID;
 import static io.dockstore.webservice.helpers.GitHubAppHelper.LAMBDA_ERROR;
 import static io.dockstore.webservice.helpers.GitHubAppHelper.handleGitHubRelease;
+import static io.openapi.api.impl.ToolClassesApiServiceImpl.COMMAND_LINE_TOOL;
+import static io.openapi.api.impl.ToolClassesApiServiceImpl.NOTEBOOK;
+import static io.openapi.api.impl.ToolClassesApiServiceImpl.WORKFLOW;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -13,6 +18,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
+import com.google.common.collect.Lists;
 import io.dockstore.client.cli.BaseIT;
 import io.dockstore.client.cli.BaseIT.TestStatus;
 import io.dockstore.client.cli.OrganizationIT;
@@ -24,21 +30,34 @@ import io.dockstore.common.SourceControl;
 import io.dockstore.common.ValidationConstants;
 import io.dockstore.openapi.client.ApiClient;
 import io.dockstore.openapi.client.ApiException;
+import io.dockstore.openapi.client.api.Ga4Ghv20Api;
 import io.dockstore.openapi.client.api.LambdaEventsApi;
+import io.dockstore.openapi.client.api.MetadataApi;
 import io.dockstore.openapi.client.api.OrganizationsApi;
 import io.dockstore.openapi.client.api.UsersApi;
 import io.dockstore.openapi.client.api.WorkflowsApi;
 import io.dockstore.openapi.client.model.Collection;
+import io.dockstore.openapi.client.model.Entry;
 import io.dockstore.openapi.client.model.LambdaEvent;
 import io.dockstore.openapi.client.model.OrcidAuthorInformation;
 import io.dockstore.openapi.client.model.Organization;
 import io.dockstore.openapi.client.model.PublishRequest;
 import io.dockstore.openapi.client.model.SourceFile;
+import io.dockstore.openapi.client.model.StarRequest;
+import io.dockstore.openapi.client.model.Tool;
+import io.dockstore.openapi.client.model.Validation;
 import io.dockstore.openapi.client.model.Workflow;
+import io.dockstore.openapi.client.model.Workflow.DescriptorTypeEnum;
+import io.dockstore.openapi.client.model.Workflow.ModeEnum;
 import io.dockstore.openapi.client.model.WorkflowSubClass;
 import io.dockstore.openapi.client.model.WorkflowVersion;
 import io.dockstore.openapi.client.model.WorkflowVersion.DescriptionSourceEnum;
+import io.dockstore.webservice.core.EntryTypeMetadata;
+import io.dockstore.webservice.helpers.GitHubAppHelper;
 import io.dockstore.webservice.helpers.GitHubAppHelper.DockstoreTestUser2Repos;
+import io.dockstore.webservice.helpers.GitHubAppHelper.DockstoreTestingRepos;
+import io.dockstore.webservice.jdbi.AppToolDAO;
+import io.dockstore.webservice.jdbi.FileDAO;
 import io.dockstore.webservice.languages.WDLHandler;
 import io.specto.hoverfly.junit.core.Hoverfly;
 import io.specto.hoverfly.junit.core.HoverflyMode;
@@ -47,8 +66,10 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import org.apache.http.HttpStatus;
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.hibernate.context.internal.ManagedSessionContext;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -72,24 +93,35 @@ class WebhookIT extends BaseIT {
     public final SystemErr systemErr = new SystemErr();
 
     private final String dockstoreTesting = "dockstore-testing";
-    private final String taggedToolRepo = "dockstore-testing/tagged-apptool";
     private final String taggedToolRepoPath = "dockstore-testing/tagged-apptool/md5sum";
     private final String workflowDockstoreYmlRepo = "dockstore-testing/workflow-dockstore-yml";
     private final String testWorkflowsAndToolsRepo = "dockstore-testing/test-workflows-and-tools";
 
+    private FileDAO fileDAO;
+    private AppToolDAO appToolDAO;
 
     @BeforeEach
-    public void cleanDB() {
+    @Override
+    public void resetDBBetweenTests() {
         CommonTestUtilities.cleanStatePrivate2(SUPPORT, false, testingPostgres);
     }
 
+    @BeforeEach
+    public void setup() {
+        DockstoreWebserviceApplication application = SUPPORT.getApplication();
+        SessionFactory sessionFactory = application.getHibernate().getSessionFactory();
+        this.fileDAO = new FileDAO(sessionFactory);
+        this.appToolDAO = new AppToolDAO(sessionFactory);
+        // used to allow us to use DAOs outside the web service
+        Session session = application.getHibernate().getSessionFactory().openSession();
+        ManagedSessionContext.bind(session);
+    }
 
     @Test
-    @Disabled("https://ucsc-cgl.atlassian.net/browse/DOCK-1890")
-    void testAppToolCollections() throws Exception {
+    void testAppToolCollections() {
         final ApiClient openApiClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
         WorkflowsApi client = new WorkflowsApi(openApiClient);
-        handleGitHubRelease(client, INSTALLATION_ID, taggedToolRepo, "refs/tags/1.0", USER_2_USERNAME);
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.TAGGED_APPTOOL, "refs/tags/1.0", USER_2_USERNAME);
         Workflow appTool = client.getWorkflowByPath("github.com/" + taggedToolRepoPath, WorkflowSubClass.APPTOOL, "versions,validations");
 
         WorkflowVersion validVersion = appTool.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).findFirst().get();
@@ -102,7 +134,7 @@ class WebhookIT extends BaseIT {
         final ApiClient webClientAdminUser = getOpenAPIWebClient(ADMIN_USERNAME, testingPostgres);
         OrganizationsApi organizationsApiAdmin = new OrganizationsApi(webClientAdminUser);
         // Create the organization
-        Organization registeredOrganization = OrganizationIT.openApiStubOrgObject();
+        Organization registeredOrganization = OrganizationIT.createOpenAPIOrg(organizationsApiAdmin);
 
         // Admin approve it
         organizationsApiAdmin.approveOrganization(registeredOrganization.getId());
@@ -115,9 +147,8 @@ class WebhookIT extends BaseIT {
         // Add tool to collection
         organizationsApiAdmin.addEntryToCollection(registeredOrganization.getId(), createdCollection.getId(), appTool.getId(), null);
 
-        // uncomment this after DOCK-1890 and delete from WebhookIT
-        // Collection collection = organizationsApiAdmin.getCollectionById(registeredOrganization.getId(), createdCollection.getId());
-        // assertTrue((collection.getEntries().stream().anyMatch(entry -> Objects.equals(entry.getId(), appTool.getId()))));
+        Collection collection = organizationsApiAdmin.getCollectionById(registeredOrganization.getId(), createdCollection.getId());
+        assertTrue((collection.getEntries().stream().anyMatch(entry -> Objects.equals(entry.getId(), appTool.getId()))));
     }
 
     @Test
@@ -250,7 +281,7 @@ class WebhookIT extends BaseIT {
 
         long userid = usersApi.getUser().getId();
 
-        handleGitHubRelease(userWorkflowsClient, INSTALLATION_ID, taggedToolRepo, "refs/tags/1.0", USER_2_USERNAME);
+        handleGitHubRelease(userWorkflowsClient, INSTALLATION_ID, DockstoreTestingRepos.TAGGED_APPTOOL, "refs/tags/1.0", USER_2_USERNAME);
 
         // Setup admin
         final ApiClient webClientAdminUser = getOpenAPIWebClient(ADMIN_USERNAME, testingPostgres);
@@ -569,12 +600,18 @@ class WebhookIT extends BaseIT {
         testDefaultVersion(client);
     }
 
-    private Workflow getFoobar1Workflow(WorkflowsApi client) {
-        return client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML + "/foobar", WorkflowSubClass.BIOWORKFLOW, "versions");
+    private Workflow getFoobar1Workflow(WorkflowsApi client, String... includes) {
+        if (includes.length == 0) {
+            return client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML + "/foobar", WorkflowSubClass.BIOWORKFLOW, "versions");
+        }
+        return client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML + "/foobar", WorkflowSubClass.BIOWORKFLOW, String.join(",", includes));
     }
 
-    private Workflow getFoobar2Workflow(WorkflowsApi client) {
-        return client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML + "/foobar2", WorkflowSubClass.BIOWORKFLOW, "versions");
+    private Workflow getFoobar2Workflow(WorkflowsApi client, String... includes) {
+        if (includes.length == 0) {
+            return client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML + "/foobar2", WorkflowSubClass.BIOWORKFLOW, "versions");
+        }
+        return client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML + "/foobar2", WorkflowSubClass.BIOWORKFLOW, String.join(",", includes));
     }
 
     /**
@@ -669,7 +706,7 @@ class WebhookIT extends BaseIT {
         final ApiClient openApiClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
         WorkflowsApi client = new WorkflowsApi(openApiClient);
 
-        handleGitHubRelease(client, INSTALLATION_ID, taggedToolRepo, "refs/tags/1.0", USER_2_USERNAME);
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.TAGGED_APPTOOL, "refs/tags/1.0", USER_2_USERNAME);
         Workflow appTool = client.getWorkflowByPath("github.com/" + taggedToolRepoPath, WorkflowSubClass.APPTOOL, "versions,validations");
 
         PublishRequest publishRequest = new PublishRequest();
@@ -786,5 +823,949 @@ class WebhookIT extends BaseIT {
             assertEquals(0, events.stream().filter(LambdaEvent::isSuccess).count(), "There should be no successful events");
             assertTrue(events.get(0).getMessage().contains(WDLHandler.ERROR_PARSING_WORKFLOW_YOU_MAY_HAVE_A_RECURSIVE_IMPORT), "Event message should indicate the problem");
         }
+    }
+
+    @Test
+    void testAppToolRSSFeedAndSiteMap() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
+        UsersApi usersApi = new UsersApi(webClient);
+        final PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+
+        // There should be no apptools
+        assertEquals(0, appToolDAO.findAllPublishedPaths().size());
+        assertEquals(0, appToolDAO.findAllPublishedPathsOrderByDbupdatedate().size());
+
+        // create and publish apptool
+        usersApi.syncUserWithGitHub();
+        GitHubAppHelper.registerAppTool(webClient);
+        handleGitHubRelease(workflowApi, INSTALLATION_ID, DockstoreTestingRepos.TAGGED_APPTOOL, "refs/tags/1.0", USER_2_USERNAME);
+        Workflow appTool = workflowApi.getWorkflowByPath("github.com/" + taggedToolRepoPath, WorkflowSubClass.APPTOOL, "versions");
+        workflowApi.publish1(appTool.getId(), publishRequest);
+
+        // There should be 1 apptool.
+        assertEquals(1, appToolDAO.findAllPublishedPaths().size());
+        assertEquals(1, appToolDAO.findAllPublishedPathsOrderByDbupdatedate().size());
+
+        final MetadataApi metadataApi = new MetadataApi(webClient);
+        String rssFeed = metadataApi.rssFeed();
+        assertTrue(rssFeed.contains("http://localhost/containers/github.com/dockstore-testing/tagged-apptool/md5sum"), "RSS feed should contain 1 apptool");
+
+        String sitemap = metadataApi.sitemap();
+        assertTrue(sitemap.contains("http://localhost/containers/github.com/dockstore-testing/tagged-apptool/md5sum"), "Sitemap with testing data should have 1 apptool");
+    }
+
+    @Test
+    void testWorkflowMigration() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
+        UsersApi usersApi = new UsersApi(webClient);
+        Workflow workflow = workflowApi
+                .manualRegister(SourceControl.GITHUB.getFriendlyName(), DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "/Dockstore.wdl",
+                        "foobar", DescriptorLanguage.WDL.getShortName(), "/test.json");
+        workflowApi.manualRegister(SourceControl.GITHUB.name(), DOCKSTORE_TEST_USER_2_HELLO_DOCKSTORE_NAME, "/Dockstore.cwl", "",
+                DescriptorLanguage.CWL.getShortName(), "/test.json");
+
+        // Refresh should work
+        workflow = workflowApi.refresh1(workflow.getId(), false);
+        assertEquals(Workflow.ModeEnum.FULL, workflow.getMode(), "Workflow should be FULL mode");
+        assertTrue(workflow.getWorkflowVersions().stream().allMatch(WorkflowVersion::isLegacyVersion), "All versions should be legacy");
+
+        // Webhook call should convert workflow to DOCKSTORE_YML
+        handleGitHubRelease(workflowApi, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/tags/0.1", USER_2_USERNAME);
+        workflow = workflowApi.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML + "/foobar", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(Workflow.ModeEnum.DOCKSTORE_YML, workflow.getMode(), "Workflow should be DOCKSTORE_YML mode");
+        assertTrue(workflow.getWorkflowVersions().stream().anyMatch(workflowVersion -> !workflowVersion.isLegacyVersion()), "One version should be not legacy");
+
+        // Refresh should now no longer work
+        try {
+            workflowApi.refresh1(workflow.getId(), false);
+            fail("Should fail on refresh and not reach this point");
+        } catch (ApiException ex) {
+            assertEquals(HttpStatus.SC_BAD_REQUEST, ex.getCode(), "Should not be able to refresh a dockstore.yml workflow.");
+        }
+
+        // Should be able to refresh a legacy version
+        workflow = workflowApi.refreshVersion(workflow.getId(), "0.2", false);
+
+        // Should not be able to refresh a GitHub App version
+        try {
+            workflowApi.refreshVersion(workflow.getId(), "0.1", false);
+            fail("Should not be able to refresh");
+        } catch (ApiException ex) {
+            assertEquals(HttpStatus.SC_BAD_REQUEST, ex.getCode());
+        }
+
+        // Refresh a version that doesn't already exist
+        try {
+            workflowApi.refreshVersion(workflow.getId(), "dne", false);
+            fail("Should not be able to refresh");
+        } catch (ApiException ex) {
+            assertEquals(HttpStatus.SC_BAD_REQUEST, ex.getCode());
+        }
+
+        List<Workflow> workflows = usersApi.addUserToDockstoreWorkflows(usersApi.getUser().getId(), "");
+        assertTrue(workflows.stream().anyMatch(wf -> Objects.equals(wf.getMode(), Workflow.ModeEnum.DOCKSTORE_YML)), "There should still be a dockstore.yml workflow");
+        assertTrue(workflows.stream().anyMatch(wf -> Objects.equals(wf.getMode(), Workflow.ModeEnum.STUB)), "There should be at least one stub workflow");
+
+        // Test that refreshing a frozen version doesn't update the version
+        testingPostgres.runUpdateStatement("UPDATE workflowversion SET commitid = NULL where name = '0.2'");
+
+        // Refresh before frozen should populate the commit id
+        workflow = workflowApi.refreshVersion(workflow.getId(), "0.2", false);
+        WorkflowVersion workflowVersion = workflow.getWorkflowVersions().stream().filter(wv -> Objects.equals(wv.getName(), "0.2")).findFirst().get();
+        assertNotNull(workflowVersion.getCommitID());
+
+        // Refresh after freezing should not update
+        testingPostgres.runUpdateStatement("UPDATE workflowversion SET commitid = NULL where name = '0.2'");
+
+        // Freeze legacy version
+        workflowVersion.setFrozen(true);
+        List<WorkflowVersion> workflowVersions = workflowApi
+                .updateWorkflowVersion(workflow.getId(), Lists.newArrayList(workflowVersion));
+        workflowVersion = workflowVersions.stream().filter(v -> v.getName().equals("0.2")).findFirst().get();
+        assertTrue(workflowVersion.isFrozen());
+
+        // Ensure refresh does not touch frozen legacy version
+        workflow = workflowApi.refreshVersion(workflow.getId(), "0.2", false);
+        assertNotNull(workflow);
+        workflowVersion = workflow.getWorkflowVersions().stream().filter(wv -> Objects.equals(wv.getName(), "0.2")).findFirst().get();
+        assertNull(workflowVersion.getCommitID());
+    }
+
+    /**
+     * This tests deleting a GitHub App workflow's default version
+     */
+    @Test
+    void testDeleteDefaultWorkflowVersion() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+        final String filterNoneWorkflowPath = "github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone";
+
+        // Add 1.0 tag and set as default version
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/tags/1.0", USER_2_USERNAME);
+        Workflow workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(1, workflow.getWorkflowVersions().size(), "should have 1 version");
+        assertNull(workflow.getDefaultVersion(), "should have no default version until set");
+        workflow = client.updateDefaultVersion1(workflow.getId(), workflow.getWorkflowVersions().get(0).getName());
+        assertNotNull(workflow.getDefaultVersion(), "should have a default version after setting");
+
+        // Add 2.0 tag
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/tags/2.0", USER_2_USERNAME);
+        workflow = client.getWorkflowByPath(filterNoneWorkflowPath, WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(2, workflow.getWorkflowVersions().size(), "should have 2 versions");
+
+        // Delete 1.0 tag, should reassign 2.0 as the default version
+        client.handleGitHubBranchDeletion(DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, USER_2_USERNAME, "refs/tags/1.0", String.valueOf(INSTALLATION_ID));
+        workflow = client.getWorkflowByPath(filterNoneWorkflowPath, WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(1, workflow.getWorkflowVersions().size(), "should have 1 version after deletion");
+        assertNotNull(workflow.getDefaultVersion(), "should have reassigned the default version during deletion");
+
+        // Publish workflow
+        PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        client.publish1(workflow.getId(), publishRequest);
+        workflow = client.getWorkflowByPath(filterNoneWorkflowPath, WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertTrue(workflow.isIsPublished());
+
+        // Delete 2.0 tag, unset default version
+        client.handleGitHubBranchDeletion(DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, USER_2_USERNAME, "refs/tags/2.0", String.valueOf(INSTALLATION_ID));
+        workflow = client.getWorkflowByPath(filterNoneWorkflowPath, WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(0, workflow.getWorkflowVersions().size(), "should have 0 versions after deletion");
+        assertNull(workflow.getDefaultVersion(), "should have no default version after final version is deleted");
+        assertFalse(workflow.isIsPublished(), "should not be published if it has 0 versions");
+    }
+
+    /**
+     * This tests calling refresh on a workflow with a Dockstore.yml
+     */
+    @Test
+    void testManualRefreshWorkflowWithGitHubApp() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        // Release 0.1 on GitHub - one new wdl workflow
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/tags/0.1", USER_2_USERNAME);
+        long workflowCount = testingPostgres.runSelectStatement("select count(*) from workflow", long.class);
+        assertEquals(1, workflowCount);
+
+        Workflow workflow = getFoobar1Workflow(client);
+        assertEquals("Apache License 2.0", workflow.getLicenseInformation().getLicenseName(), "Should be able to get license after GitHub App register");
+
+        // Ensure that new workflow is created and is what is expected
+
+        assertEquals(Workflow.DescriptorTypeEnum.WDL, workflow.getDescriptorType(), "Should be a WDL workflow");
+        assertEquals(Workflow.ModeEnum.DOCKSTORE_YML, workflow.getMode(), "Should be type DOCKSTORE_YML");
+        assertTrue(workflow.getWorkflowVersions().stream().anyMatch((WorkflowVersion version) -> Objects.equals(version.getName(), "0.1")), "Should have a 0.1 version.");
+        boolean hasLegacyVersion = workflow.getWorkflowVersions().stream().anyMatch(WorkflowVersion::isLegacyVersion);
+        assertFalse(hasLegacyVersion, "Workflow should not have any legacy refresh versions.");
+
+        // Refresh
+        try {
+            client.refresh1(workflow.getId(), false);
+            fail("Should fail on refresh and not reach this point");
+        } catch (ApiException ex) {
+            assertEquals(HttpStatus.SC_BAD_REQUEST, ex.getCode(), "Should not be able to refresh a dockstore.yml workflow.");
+        }
+    }
+
+    /**
+     * This tests the GitHub release process does not work for users that do not exist on Dockstore
+     */
+    @Test
+    void testGitHubReleaseNoWorkflowOnDockstoreNoUser() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        try {
+            handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/tags/0.1", "thisisafakeuser");
+            fail("Should not reach this statement.");
+        } catch (ApiException ex) {
+            assertEquals(LAMBDA_ERROR, ex.getCode(), "Should not be able to add a workflow when user does not exist on Dockstore.");
+        }
+    }
+
+    /**
+     * Tests:
+     * An unpublished workflow with invalid versions can have its descriptor type changed
+     * The workflow can then have new valid versions registered
+     * The valid workflow cannot have its descriptor type changed anymore (because it's valid)
+     * The published workflow cannot have its descriptor type changed anymore (because it's published)
+     */
+    @Test
+    void testDescriptorChange() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi workflowsApi = new WorkflowsApi(webClient);
+        handleGitHubRelease(workflowsApi, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/heads/missingPrimaryDescriptor", USER_2_USERNAME);
+        long workflowCount = testingPostgres.runSelectStatement("select count(*) from workflow", long.class);
+        assertEquals(1, workflowCount);
+        // Ensure that new workflow is created and is what is expected
+        Workflow workflow = getFoobar1Workflow(workflowsApi);
+        assertEquals(Workflow.DescriptorTypeEnum.WDL, workflow.getDescriptorType(), "Should be a WDL workflow");
+        assertEquals(Workflow.ModeEnum.DOCKSTORE_YML, workflow.getMode(), "Should be type DOCKSTORE_YML");
+        assertEquals(1, workflow.getWorkflowVersions().size(), "Should have one version");
+        assertFalse(workflow.getWorkflowVersions().get(0).isValid(), "Should be invalid (wrong language, bad version)");
+
+        workflowsApi.updateDescriptorType(workflow.getId(), DescriptorLanguage.CWL.toString());
+        Workflow updatedWorkflowAfterModifyingDescriptorType = workflowsApi.getWorkflow(workflow.getId(), "");
+        assertEquals(Workflow.DescriptorTypeEnum.CWL, updatedWorkflowAfterModifyingDescriptorType.getDescriptorType(),
+            "The descriptor language should have been changed");
+        assertEquals(0, updatedWorkflowAfterModifyingDescriptorType.getWorkflowVersions().size(), "The old versions should have been removed");
+
+        workflowsApi.updateDescriptorType(workflow.getId(), DescriptorLanguage.WDL.toString());
+        updatedWorkflowAfterModifyingDescriptorType = workflowsApi.getWorkflow(workflow.getId(), "versions");
+        assertEquals(Workflow.DescriptorTypeEnum.WDL, updatedWorkflowAfterModifyingDescriptorType.getDescriptorType(),
+            "The descriptor language should have been changed");
+        assertEquals(0, updatedWorkflowAfterModifyingDescriptorType.getWorkflowVersions().size(), "The old versions should have been removed");
+
+        // Release 0.1 on GitHub - one new wdl workflow
+        handleGitHubRelease(workflowsApi, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/tags/0.1", USER_2_USERNAME);
+        workflowCount = testingPostgres.runSelectStatement("select count(*) from workflow", long.class);
+        assertEquals(1, workflowCount);
+
+        // Ensure that new workflow is created and is what is expected
+        workflow = getFoobar1Workflow(workflowsApi);
+        assertEquals(Workflow.DescriptorTypeEnum.WDL, workflow.getDescriptorType(), "Should be a WDL workflow");
+        assertEquals(Workflow.ModeEnum.DOCKSTORE_YML, workflow.getMode(), "Should be type DOCKSTORE_YML");
+        assertEquals(1, workflow.getWorkflowVersions().size(), "Should have one version 0.1");
+        assertTrue(workflow.getWorkflowVersions().get(0).isValid(), "Should be valid");
+        try {
+            workflowsApi
+                    .updateDescriptorType(workflow.getId(), DescriptorLanguage.CWL.toString());
+            fail("Should not be able to change the descriptor type of a workflow that has valid versions");
+        } catch (ApiException e) {
+            assertEquals("Cannot change descriptor type of a valid workflow", e.getMessage());
+        }
+        PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        workflowsApi.publish1(workflow.getId(), publishRequest);
+        try {
+            workflowsApi.updateDescriptorType(workflow.getId(), DescriptorLanguage.WDL.toString());
+            fail("Should also not be able to change the descriptor type of a workflow that is published");
+        } catch (ApiException e) {
+            assertEquals("Cannot change descriptor type of a published workflow", e.getMessage());
+        }
+    }
+
+    /**
+     * This tests the GitHub release process when the dockstore.yml is
+     * * Missing the primary descriptor
+     * * Missing a test parameter file
+     * * Has an unknown property
+     */
+    @Test
+    void testInvalidDockstoreYmlFiles() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+        UsersApi usersApi = new UsersApi(webClient);
+
+        // Release 0.1 on GitHub - one new wdl workflow
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/tags/0.1", USER_2_USERNAME);
+        long workflowCount = testingPostgres.runSelectStatement("select count(*) from workflow", long.class);
+        assertEquals(1, workflowCount);
+
+        // Ensure that new workflow is created and is what is expected
+        Workflow workflow = getFoobar1Workflow(client);
+        assertEquals(DescriptorTypeEnum.WDL, workflow.getDescriptorType(), "Should be a WDL workflow");
+        assertEquals(ModeEnum.DOCKSTORE_YML, workflow.getMode(), "Should be type DOCKSTORE_YML");
+        assertEquals(1, workflow.getWorkflowVersions().size(), "Should have one version 0.1");
+        assertTrue(workflow.getWorkflowVersions().get(0).isValid(), "Should be valid");
+        assertNull(getLatestLambdaEventMessage(0, usersApi), "Lambda event message should be empty");
+
+        // Push missingPrimaryDescriptor on GitHub - one existing wdl workflow, missing primary descriptor
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/heads/missingPrimaryDescriptor", USER_2_USERNAME);
+        workflowCount = testingPostgres.runSelectStatement("select count(*) from workflow", long.class);
+        assertEquals(1, workflowCount);
+
+        // Ensure that new version is in the correct state (invalid)
+        workflow = getFoobar1Workflow(client, "validations");
+        assertNotNull(workflow);
+        assertEquals(2, workflow.getWorkflowVersions().size(), "Should have two versions");
+
+        WorkflowVersion missingPrimaryDescriptorVersion = workflow.getWorkflowVersions().stream().filter(workflowVersion -> Objects.equals(workflowVersion.getName(), "missingPrimaryDescriptor")).findFirst().get();
+        assertFalse(missingPrimaryDescriptorVersion.isValid(), "Version should be invalid");
+
+        // Check existence of files and validations
+        List<io.dockstore.webservice.core.SourceFile> sourceFiles = fileDAO.findSourceFilesByVersion(missingPrimaryDescriptorVersion.getId());
+        assertTrue(sourceFiles.stream().anyMatch(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), DOCKSTORE_YML_PATH)), "Should have .dockstore.yml file");
+        assertTrue(sourceFiles.stream().filter(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), "/doesnotexist.wdl")).findFirst().isEmpty(),
+            "Should not have doesnotexist.wdl file");
+        assertFalse(missingPrimaryDescriptorVersion.getValidations().stream().filter(validation -> Objects.equals(validation.getType(), Validation.TypeEnum.DOCKSTORE_YML)).findFirst().get().isValid(),
+            "Should have invalid .dockstore.yml");
+        assertFalse(missingPrimaryDescriptorVersion.getValidations().stream().filter(validation -> Objects.equals(validation.getType(), Validation.TypeEnum.DOCKSTORE_WDL)).findFirst().get().isValid(),
+            "Should have invalid doesnotexist.wdl");
+        assertTrue(getLatestLambdaEventMessage(0, usersApi).contains("descriptor"), "Refers to missing primary descriptor");
+
+        // Push missingTestParameterFile on GitHub - one existing wdl workflow, missing a test parameter file
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/heads/missingTestParameterFile", USER_2_USERNAME);
+        workflowCount = testingPostgres.runSelectStatement("select count(*) from workflow", long.class);
+        assertEquals(1, workflowCount);
+
+        // Ensure that new version is in the correct state (invalid)
+        workflow = getFoobar1Workflow(client, "validations");
+        assertNotNull(workflow);
+        assertEquals(3, workflow.getWorkflowVersions().size(), "Should have three versions");
+
+        WorkflowVersion missingTestParameterFileVersion = workflow.getWorkflowVersions().stream().filter(workflowVersion -> Objects.equals(workflowVersion.getName(), "missingTestParameterFile")).findFirst().get();
+        assertTrue(missingTestParameterFileVersion.isValid(), "Version should be valid (missing test parameter doesn't make the version invalid)");
+
+        // Check existence of files and validations
+        sourceFiles = fileDAO.findSourceFilesByVersion(missingTestParameterFileVersion.getId());
+        assertTrue(sourceFiles.stream().anyMatch(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), DOCKSTORE_YML_PATH)), "Should have .dockstore.yml file");
+        assertTrue(sourceFiles.stream().filter(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), "/test/doesnotexist.txt")).findFirst().isEmpty(),
+            "Should not have /test/doesnotexist.txt file");
+        assertTrue(sourceFiles.stream().anyMatch(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), "/Dockstore2.wdl")), "Should have Dockstore2.wdl file");
+        assertFalse(missingTestParameterFileVersion.getValidations().stream().filter(validation -> Objects.equals(validation.getType(), Validation.TypeEnum.DOCKSTORE_YML)).findFirst().get().isValid(),
+            "Should have invalid .dockstore.yml");
+        assertTrue(missingTestParameterFileVersion.getValidations().stream().filter(validation -> Objects.equals(validation.getType(), Validation.TypeEnum.DOCKSTORE_WDL)).findFirst().get().isValid(),
+            "Should have valid Dockstore2.wdl");
+        assertTrue(getLatestLambdaEventMessage(0, usersApi).contains("/idonotexist.json"), "Refers to missing test file");
+
+        // Push unknownProperty on GitHub - one existing wdl workflow, incorrectly spelled testParameterFiles property
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/heads/unknownProperty", USER_2_USERNAME);
+        workflowCount = testingPostgres.runSelectStatement("select count(*) from workflow", long.class);
+        assertEquals(1, workflowCount);
+
+        // Ensure that new version is in the correct state (valid)
+        workflow = getFoobar1Workflow(client, "validations");
+        assertNotNull(workflow);
+        assertEquals(4, workflow.getWorkflowVersions().size(), "Should have four versions");
+
+        WorkflowVersion unknownPropertyVersion = workflow.getWorkflowVersions().stream().filter(workflowVersion -> Objects.equals(workflowVersion.getName(), "unknownProperty")).findFirst().get();
+        assertTrue(unknownPropertyVersion.isValid(), "Version should be valid (unknown property doesn't make the version invalid)");
+
+        // Check existence of files and validations
+        sourceFiles = fileDAO.findSourceFilesByVersion(unknownPropertyVersion.getId());
+        assertTrue(sourceFiles.stream().anyMatch(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), DOCKSTORE_YML_PATH)), "Should have .dockstore.yml file");
+        assertTrue(sourceFiles.stream().filter(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), "/dockstore.wdl.json")).findFirst().isEmpty(),
+            "Should not have /dockstore.wdl.json file");
+        assertTrue(sourceFiles.stream().anyMatch(sourceFile -> Objects.equals(sourceFile.getAbsolutePath(), "/Dockstore2.wdl")), "Should have Dockstore2.wdl file");
+        assertFalse(missingTestParameterFileVersion.getValidations().stream().filter(validation -> Objects.equals(validation.getType(), Validation.TypeEnum.DOCKSTORE_YML)).findFirst().get().isValid(),
+            "Should have invalid .dockstore.yml");
+        assertTrue(missingTestParameterFileVersion.getValidations().stream().filter(validation -> Objects.equals(validation.getType(), Validation.TypeEnum.DOCKSTORE_WDL)).findFirst().get().isValid(),
+            "Should have valid Dockstore2.wdl");
+        assertTrue(getLatestLambdaEventMessage(0, usersApi).contains("testParameterFilets"), "Refers to misspelled property");
+
+        // There should be 4 successful lambda events
+        List<LambdaEvent> events = usersApi.getUserGitHubEvents(0, 10);
+        assertEquals(4, events.stream().filter(LambdaEvent::isSuccess).count(), "There should be 4 successful events");
+
+        final int versionCountBeforeInvalidDockstoreYml = getFoobar1Workflow(client).getWorkflowVersions().size();
+        // Push branch with invalid dockstore.yml
+        try {
+            handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/heads/invalidDockstoreYml", USER_2_USERNAME);
+            fail("Should not reach this statement");
+        } catch (ApiException ex) {
+            List<LambdaEvent> failEvents = usersApi.getUserGitHubEvents(0, 10);
+            assertEquals(1, failEvents.stream().filter(lambdaEvent -> !lambdaEvent.isSuccess()).count(), "There should be 1 unsuccessful event");
+            assertEquals(versionCountBeforeInvalidDockstoreYml, getFoobar1Workflow(client).getWorkflowVersions().size(), "Number of versions should be the same");
+        }
+    }
+
+    private LambdaEvent getLatestLambdaEvent(Integer offset, UsersApi usersApi) {
+        return usersApi.getUserGitHubEvents(offset, 1).get(0);
+    }
+
+    private String getLatestLambdaEventMessage(Integer offset, UsersApi usersApi) {
+        return getLatestLambdaEvent(offset, usersApi).getMessage();
+    }
+
+    /**
+     * Test that a .dockstore.yml workflow has the expected path for its test parameter file.
+     */
+    @Test
+    void testTestParameterPaths() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.WORKFLOW_DOCKSTORE_YML, "refs/heads/master", USER_2_USERNAME);
+        Workflow workflow = getFoobar1Workflow(client);
+        WorkflowVersion version = workflow.getWorkflowVersions().get(0);
+        List<io.dockstore.webservice.core.SourceFile> sourceFiles = fileDAO.findSourceFilesByVersion(version.getId());
+        assertTrue(sourceFiles.stream().anyMatch(sourceFile -> sourceFile.getPath().equals("/dockstore.wdl.json")), "Test file should have the expected path");
+    }
+
+    /**
+     * This tests the GitHub release with .dockstore.yml located in /.github/.dockstore.yml
+     */
+    @Test
+    void testGithubDirDockstoreYml() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/tags/1.0", USER_2_USERNAME);
+        Workflow workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "");
+        assertNotNull(workflow);
+    }
+
+    /**
+     * This tests filters functionality in .dockstore.yml
+     * <a href="https://github.com/DockstoreTestUser2/dockstoreyml-github-filters-test">...</a>
+     * Workflow filters are configured as follows:
+     * * filterbranch filters for "develop"
+     * * filtertag filters for "1.0"
+     * * filtermulti filters for "dev*" and "1.*"
+     * * filternone has no filters (accepts all tags & branches)
+     * * filterregexerror has a filter with an invalid regex string (matches nothing)
+     */
+    @Test
+    void testDockstoreYmlFilters() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        // master should be excluded by all of the workflows with filters
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/heads/master", USER_2_USERNAME);
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterbranch", WorkflowSubClass.BIOWORKFLOW, ""));
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtertag", WorkflowSubClass.BIOWORKFLOW, ""));
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtermulti", WorkflowSubClass.BIOWORKFLOW, ""));
+        Workflow workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(1, workflow.getWorkflowVersions().size());
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterregexerror", WorkflowSubClass.BIOWORKFLOW, ""));
+
+        // tag 2.0 should be excluded by all of the workflows with filters
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/tags/2.0", USER_2_USERNAME);
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterbranch", WorkflowSubClass.BIOWORKFLOW, ""));
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtertag", WorkflowSubClass.BIOWORKFLOW, ""));
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtermulti", WorkflowSubClass.BIOWORKFLOW, ""));
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(2, workflow.getWorkflowVersions().size());
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterregexerror", WorkflowSubClass.BIOWORKFLOW, ""));
+
+        // develop2 should be accepted by the heads/dev* filter in filtermulti
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/heads/develop2", USER_2_USERNAME);
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterbranch", WorkflowSubClass.BIOWORKFLOW, ""));
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtertag", WorkflowSubClass.BIOWORKFLOW, ""));
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtermulti", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(1, workflow.getWorkflowVersions().size());
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(3, workflow.getWorkflowVersions().size());
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterregexerror", WorkflowSubClass.BIOWORKFLOW, ""));
+
+        // tag 1.1 should be accepted by the 1.* filter in filtermulti
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/tags/1.1", USER_2_USERNAME);
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterbranch", WorkflowSubClass.BIOWORKFLOW, ""));
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtertag", WorkflowSubClass.BIOWORKFLOW, ""));
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtermulti", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(2, workflow.getWorkflowVersions().size());
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(4, workflow.getWorkflowVersions().size());
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterregexerror", WorkflowSubClass.BIOWORKFLOW, ""));
+
+        // tag 1.0 should be accepted by tags/1.0 in filtertag and 1.* in filtermulti
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/tags/1.0", USER_2_USERNAME);
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterbranch", WorkflowSubClass.BIOWORKFLOW, ""));
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtertag", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(1, workflow.getWorkflowVersions().size());
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtermulti", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(3, workflow.getWorkflowVersions().size());
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(5, workflow.getWorkflowVersions().size());
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterregexerror", WorkflowSubClass.BIOWORKFLOW, ""));
+
+        // develop should be accepted by develop in filterbranch and heads/dev* in filtermulti
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/heads/develop", USER_2_USERNAME);
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterbranch", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(1, workflow.getWorkflowVersions().size());
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtertag", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(1, workflow.getWorkflowVersions().size());
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filtermulti", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(4, workflow.getWorkflowVersions().size());
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "versions");
+        assertEquals(6, workflow.getWorkflowVersions().size());
+        assertThrows(ApiException.class, () -> client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filterregexerror", WorkflowSubClass.BIOWORKFLOW, ""));
+    }
+
+    /**
+     * This tests publishing functionality in .dockstore.yml
+     */
+    @Test
+    void testDockstoreYmlPublish() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/heads/publish", USER_2_USERNAME);
+        Workflow workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "");
+        assertEquals(1, testingPostgres.getPublishEventCountForWorkflow(workflow.getId()));
+        assertEquals(0, testingPostgres.getUnpublishEventCountForWorkflow(workflow.getId()));
+        assertTrue(workflow.isIsPublished());
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/heads/unpublish", USER_2_USERNAME);
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST + "/filternone", WorkflowSubClass.BIOWORKFLOW, "");
+        assertFalse(workflow.isIsPublished());
+        assertEquals(1, testingPostgres.getPublishEventCountForWorkflow(workflow.getId()));
+        assertEquals(1, testingPostgres.getUnpublishEventCountForWorkflow(workflow.getId()));
+    }
+
+    /**
+     * This tests multiple authors functionality in .dockstore.yml and descriptor file.
+     * If there are authors in .dockstore.yml, then only .dockstore.yml authors are saved, even if the descriptor has an author.
+     * If there are no authors in .dockstore.yml, then authors from the descriptor are saved.
+     */
+    @Test
+    void testDockstoreYmlAuthors() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+        String wdlWorkflowRepoPath = String.format("github.com/%s/%s", DockstoreTestUser2Repos.TEST_AUTHORS, "foobar");
+        String cwlWorkflowRepoPath = String.format("github.com/%s/%s", DockstoreTestUser2Repos.TEST_AUTHORS, "foobar2");
+        String nextflowWorkflowRepoPath = String.format("github.com/%s/%s", DockstoreTestUser2Repos.TEST_AUTHORS, "foobar3");
+        String wdl2WorkflowRepoPath = String.format("github.com/%s/%s", DockstoreTestUser2Repos.TEST_AUTHORS, "foobar4");
+        Workflow workflow;
+        WorkflowVersion version;
+
+        // Workflows containing 1 descriptor author and multiple .dockstore.yml authors.
+        // If the .dockstore.yml specifies an author, then only the .dockstore.yml's authors should be saved
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_AUTHORS, "refs/heads/main", USER_2_USERNAME);
+        // WDL workflow
+        workflow = client.getWorkflowByPath(wdlWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("main")).findFirst().get();
+        assertEquals(2, version.getAuthors().size());
+        assertEquals(2, version.getOrcidAuthors().size());
+        final String wdlDescriptorAuthorName = "Descriptor Author";
+        assertTrue(version.getAuthors().stream().noneMatch(author -> author.getName().equals(wdlDescriptorAuthorName)), "Should not have any author from the descriptor");
+        // CWL workflow
+        workflow = client.getWorkflowByPath(cwlWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("main")).findFirst().get();
+        assertEquals(1, version.getAuthors().size());
+        assertEquals(1, version.getOrcidAuthors().size());
+        final String cwlDescriptorAuthorName = "Test User";
+        assertTrue(version.getAuthors().stream().noneMatch(author -> author.getName().equals(cwlDescriptorAuthorName)), "Should not have any author from the descriptor");
+        // Nextflow workflow
+        workflow = client.getWorkflowByPath(nextflowWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("main")).findFirst().get();
+        assertEquals(1, version.getAuthors().size());
+        assertEquals(1, version.getOrcidAuthors().size());
+        final String nextflowDescriptorAuthorName = "Nextflow Test Author";
+        assertTrue(version.getAuthors().stream().noneMatch(author -> author.getName().equals(nextflowDescriptorAuthorName)), "Should not have any author from the descriptor");
+        // WDL workflow containing 1 descriptor author, 1 ORCID author, and 0 non-ORCID authors
+        workflow = client.getWorkflowByPath(wdl2WorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("main")).findFirst().get();
+        assertEquals(0, version.getAuthors().size());
+        assertEquals(1, version.getOrcidAuthors().size());
+        assertTrue(version.getAuthors().stream().noneMatch(author -> author.getName().equals(wdlDescriptorAuthorName)), "Should not have any author from the descriptor");
+
+        // WDL workflow containing only .dockstore.yml authors
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_AUTHORS, "refs/heads/onlyDockstoreYmlAuthors", USER_2_USERNAME);
+        workflow = client.getWorkflowByPath(wdlWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("onlyDockstoreYmlAuthors")).findFirst().get();
+        assertEquals(2, version.getAuthors().size());
+        assertEquals(2, version.getOrcidAuthors().size());
+
+        // WDL workflow containing only a descriptor author
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_AUTHORS, "refs/heads/onlyDescriptorAuthor", USER_2_USERNAME);
+        workflow = client.getWorkflowByPath(wdlWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("onlyDescriptorAuthor")).findFirst().get();
+        assertEquals(1, version.getAuthors().size());
+        assertEquals(wdlDescriptorAuthorName, version.getAuthor());
+        assertEquals(0, version.getOrcidAuthors().size());
+
+        // Release WDL workflow containing only a descriptor author again and test that it doesn't create a duplicate author
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_AUTHORS, "refs/heads/onlyDescriptorAuthor", USER_2_USERNAME);
+        workflow = client.getWorkflowByPath(wdlWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("onlyDescriptorAuthor")).findFirst().get();
+        assertEquals(1, version.getAuthors().size());
+        assertEquals(wdlDescriptorAuthorName, version.getAuthor());
+        assertEquals(0, version.getOrcidAuthors().size());
+
+        // WDL workflow containing multiple descriptor authors separated by a comma ("Author 1, Author 2") and no .dockstore.yml authors
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_AUTHORS, "refs/heads/multipleDescriptorAuthors", USER_2_USERNAME);
+        workflow = client.getWorkflowByPath(wdlWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
+        version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("multipleDescriptorAuthors")).findFirst().get();
+        assertEquals(2, version.getAuthors().size());
+        version.getAuthors().forEach(author -> assertNotNull(author.getEmail()));
+        assertEquals(0, version.getOrcidAuthors().size());
+    }
+
+    // .dockstore.yml in test repo needs to change to add a 'name' field to one of them. Should also include another branch that doesn't keep the name field
+    @Test
+    void testTools() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        UsersApi usersApi = new UsersApi(webClient);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/main", USER_2_USERNAME);
+        Workflow appTool = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS_TOOL_PATH, WorkflowSubClass.APPTOOL, "versions");
+        Workflow workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, WorkflowSubClass.BIOWORKFLOW, "versions");
+
+        assertNotNull(workflow);
+        assertNotNull(appTool);
+
+        assertEquals(1, appTool.getWorkflowVersions().size());
+        assertEquals(1, workflow.getWorkflowVersions().size());
+
+        Long userId = usersApi.getUser().getId();
+        List<io.dockstore.openapi.client.model.Workflow> usersAppTools = usersApi.userAppTools(userId);
+        assertEquals(1, usersAppTools.size());
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/invalid-workflow", USER_2_USERNAME);
+        appTool = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS_TOOL_PATH, WorkflowSubClass.APPTOOL, "versions,validations");
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, WorkflowSubClass.BIOWORKFLOW, "versions,validations");
+        assertEquals(2, appTool.getWorkflowVersions().size());
+        assertEquals(2, workflow.getWorkflowVersions().size());
+
+        WorkflowVersion invalidVersion = workflow.getWorkflowVersions().stream().filter(workflowVersion -> !workflowVersion.isValid()).findFirst().get();
+        assertFalse(invalidVersion.getValidations().isEmpty());
+        Validation workflowValidation = invalidVersion.getValidations().stream().filter(validation -> validation.getType().equals(Validation.TypeEnum.DOCKSTORE_CWL)).findFirst().get();
+        assertFalse(workflowValidation.isValid());
+        assertTrue(workflowValidation.getMessage().contains("Did you mean to register a tool"));
+        appTool.getWorkflowVersions().forEach(workflowVersion -> {
+            if (!workflowVersion.isValid()) {
+                fail("Tool should be valid for both versions");
+            }
+        });
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/invalidTool", USER_2_USERNAME);
+        appTool = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS_TOOL_PATH, WorkflowSubClass.APPTOOL, "versions,validations");
+        workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, WorkflowSubClass.BIOWORKFLOW, "versions,validations");
+        assertEquals(3, appTool.getWorkflowVersions().size());
+        assertEquals(3, workflow.getWorkflowVersions().size());
+
+        invalidVersion = appTool.getWorkflowVersions().stream().filter(workflowVersion -> !workflowVersion.isValid()).findFirst().get();
+        Validation toolValidation = invalidVersion.getValidations().stream().filter(validation -> validation.getType().equals(Validation.TypeEnum.DOCKSTORE_CWL)).findFirst().get();
+        assertFalse(toolValidation.isValid());
+        assertTrue(toolValidation.getMessage().contains("Did you mean to register a workflow"));
+
+        // publish endpoint updates elasticsearch index
+        PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        WorkflowVersion validVersion = appTool.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).findFirst().get();
+        testingPostgres.runUpdateStatement("update apptool set actualdefaultversion = " + validVersion.getId() + " where id = " + appTool.getId());
+        client.publish1(appTool.getId(), publishRequest);
+        client.publish1(workflow.getId(), publishRequest);
+        assertFalse(systemOut.getText().contains("Could not submit index to elastic search"));
+
+        Ga4Ghv20Api ga4Ghv20Api = new Ga4Ghv20Api(webClient);
+        final List<io.dockstore.openapi.client.model.Tool> tools = ga4Ghv20Api.toolsGet(null, null, null, null, null, null, null, null, null, null, null, null, null);
+        assertEquals(2, tools.size());
+
+        final io.dockstore.openapi.client.model.Tool tool = ga4Ghv20Api.toolsIdGet("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS_TOOL_PATH);
+        assertNotNull(tool);
+        assertEquals("CommandLineTool", tool.getToolclass().getDescription());
+        final Tool trsWorkflow = ga4Ghv20Api.toolsIdGet(EntryTypeMetadata.WORKFLOW.getTrsPrefix() + "/github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS);
+        assertNotNull(trsWorkflow);
+        assertEquals("Workflow", trsWorkflow.getToolclass().getDescription());
+
+        publishRequest.setPublish(false);
+        client.publish1(appTool.getId(), publishRequest);
+        client.publish1(workflow.getId(), publishRequest);
+        assertFalse(systemOut.getText().contains("Could not submit index to elastic search"));
+    }
+
+    @Test
+    void testSnapshotAppTool() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.TAGGED_APPTOOL, "refs/tags/1.0", USER_2_USERNAME);
+        Workflow appTool = client.getWorkflowByPath("github.com/" + taggedToolRepoPath, WorkflowSubClass.APPTOOL, "versions,validations");
+
+        PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        WorkflowVersion validVersion = appTool.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).findFirst().get();
+        testingPostgres.runUpdateStatement("update apptool set actualdefaultversion = " + validVersion.getId() + " where id = " + appTool.getId());
+        client.publish1(appTool.getId(), publishRequest);
+
+        // snapshot the version
+        validVersion.setFrozen(true);
+        client.updateWorkflowVersion(appTool.getId(), Lists.newArrayList(validVersion));
+
+        // check if version is frozen
+        appTool = client.getWorkflow(appTool.getId(), null);
+        validVersion = appTool.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).findFirst().get();
+        assertTrue(validVersion.isFrozen());
+
+        // check if image has been created
+        long imageCount = testingPostgres.runSelectStatement("select count(*) from entry_version_image where versionid = " + validVersion.getId(), long.class);
+        assertEquals(1, imageCount);
+    }
+
+    @Test
+    void testChangingAppToolTopics() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.TAGGED_APPTOOL, "refs/tags/1.0", USER_2_USERNAME);
+        Workflow appTool = client.getWorkflowByPath("github.com/" + taggedToolRepoPath, WorkflowSubClass.APPTOOL, "versions,validations");
+
+        PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        client.publish1(appTool.getId(), publishRequest);
+
+        String newTopic = "this is a new topic";
+        appTool.setTopicManual(newTopic);
+        appTool = client.updateWorkflow(appTool.getId(), appTool);
+        assertEquals(newTopic, appTool.getTopicManual());
+    }
+
+    @Test
+    void testStarAppTool() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        UsersApi usersApi = new UsersApi(webClient);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/main", USER_2_USERNAME);
+        Workflow appTool = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS_TOOL_PATH, WorkflowSubClass.APPTOOL, "versions,validations");
+
+        PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        WorkflowVersion validVersion = appTool.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).findFirst().get();
+        testingPostgres.runUpdateStatement("update apptool set actualdefaultversion = " + validVersion.getId() + " where id = " + appTool.getId());
+        client.publish1(appTool.getId(), publishRequest);
+
+        List<io.dockstore.openapi.client.model.Entry> pre = usersApi.getStarredTools();
+        assertEquals(0, pre.stream().filter(e -> e.getId().equals(appTool.getId())).count());
+        assertEquals(0, client.getStarredUsers1(appTool.getId()).size());
+
+        client.starEntry1(appTool.getId(), new StarRequest().star(true));
+
+        List<Entry> post = usersApi.getStarredTools();
+        assertEquals(1, post.stream().filter(e -> e.getId().equals(appTool.getId())).count());
+        assertEquals(pre.size() + 1, post.size());
+        assertEquals(1, client.getStarredUsers1(appTool.getId()).size());
+    }
+
+    @Test
+    void testTRSWithAppTools() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/main", USER_2_USERNAME);
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/invalid-workflow", USER_2_USERNAME);
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/invalidTool", USER_2_USERNAME);
+        Workflow appTool = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS_TOOL_PATH, WorkflowSubClass.APPTOOL, "versions,validations");
+        Workflow workflow = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, WorkflowSubClass.BIOWORKFLOW, "versions,validations");        // publish endpoint updates elasticsearch index
+        PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
+        WorkflowVersion validVersion = appTool.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).findFirst().get();
+        testingPostgres.runUpdateStatement("update apptool set actualdefaultversion = " + validVersion.getId() + " where id = " + appTool.getId());
+        client.publish1(appTool.getId(), publishRequest);
+        client.publish1(workflow.getId(), publishRequest);
+
+
+        Ga4Ghv20Api ga4Ghv20Api = new Ga4Ghv20Api(webClient);
+        List<io.dockstore.openapi.client.model.Tool> tools = ga4Ghv20Api.toolsGet(null, null, null, null, null, null, null, null, null, null, null, null, null);
+        assertEquals(2, tools.size());
+
+        // testing filters of various kinds
+
+        tools = ga4Ghv20Api.toolsGet(null, null, null, null, null, null, null, null, null, null, true, null, null);
+        // neither the apptool or the regular workflow are checkers
+        assertEquals(0, tools.size());
+        tools = ga4Ghv20Api.toolsGet(null, null, null, null, null, null, null, null, null, null, false, null, null);
+        // neither the apptool or the regular workflow are checkers
+        assertEquals(2, tools.size());
+        tools = ga4Ghv20Api.toolsGet(null, null, WORKFLOW, null, null, null, null, null, null, null, false, null, null);
+        // the apptool is a commandline tool and not a workflow
+        assertEquals(1, tools.size());
+        tools = ga4Ghv20Api.toolsGet(null, null, COMMAND_LINE_TOOL, null, null, null, null, null, null, null, false, null, null);
+        // the apptool is a commandline tool and not a workflow
+        assertEquals(1, tools.size());
+        tools = ga4Ghv20Api.toolsGet(null, null, SERVICE, null, null, null, null, null, null, null, false, null, null);
+        // neither are services
+        assertEquals(0, tools.size());
+        tools = ga4Ghv20Api.toolsGet(null, null, null, DescriptorLanguage.SERVICE.getShortName(), null, null, null, null, null, null, false, null, null);
+        // neither are services this way either
+        assertEquals(0, tools.size());
+        tools = ga4Ghv20Api.toolsGet(null, null, NOTEBOOK, null, null, null, null, null, null, null, false, null, null);
+        // no notebooks
+        assertEquals(0, tools.size());
+
+        // testing paging
+
+        tools = ga4Ghv20Api.toolsGet(null, null, null, DescriptorLanguage.CWL.getShortName(), null, null, null, null, null, null, false, String.valueOf(-1), 1);
+        // should just go to first page
+        assertEquals(1, tools.size());
+        assertEquals(WORKFLOW, tools.get(0).getToolclass().getDescription());
+        tools = ga4Ghv20Api.toolsGet(null, null, null, DescriptorLanguage.CWL.getShortName(), null, null, null, null, null, null, false, String.valueOf(0), 1);
+        // first page
+        assertEquals(1, tools.size());
+        assertEquals(WORKFLOW, tools.get(0).getToolclass().getDescription());
+        tools = ga4Ghv20Api.toolsGet(null, null, null, DescriptorLanguage.CWL.getShortName(), null, null, null, null, null, null, false, String.valueOf(1), 1);
+        // second page
+        assertEquals(1, tools.size());
+        assertEquals(COMMAND_LINE_TOOL, tools.get(0).getToolclass().getDescription());
+        tools = ga4Ghv20Api.toolsGet(null, null, null, DescriptorLanguage.CWL.getShortName(), null, null, null, null, null, null, false, String.valueOf(1000), 1);
+        //TODO should just go to second page, but for now I guess you just scroll off into nothingness
+        assertEquals(0, tools.size());
+    }
+
+    @Test
+    void testDuplicatePathsAcrossTables() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        try {
+            handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.TEST_WORKFLOW_AND_TOOLS, "refs/heads/duplicate-paths", USER_2_USERNAME);
+            fail("Should not be able to create a workflow and apptool with the same path.");
+        } catch (ApiException ex) {
+            assertTrue(ex.getMessage().contains("have no name"));
+        }
+
+        // Check that the database trigger created an entry in fullworkflowpath table
+        long pathCount = testingPostgres.runSelectStatement("select count(*) from fullworkflowpath", long.class);
+        assertEquals(0, pathCount);
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestUser2Repos.DOCKSTOREYML_GITHUB_FILTERS_TEST, "refs/tags/1.0", USER_2_USERNAME);
+        pathCount = testingPostgres.runSelectStatement("select count(*) from fullworkflowpath", long.class);
+        assertTrue(pathCount >= 3);
+
+        try {
+            testingPostgres.runUpdateStatement("INSERT INTO fullworkflowpath(id, organization, repository, sourcecontrol, workflowname) VALUES (1010, 'DockstoreTestUser2', 'dockstoreyml-github-filters-test', 'github.com', 'filternone')");
+            fail("Database should prevent duplicate paths between tables");
+        } catch (Exception ex) {
+            assertTrue(ex.getMessage().contains("duplicate key value violates"));
+        }
+    }
+
+    private long countTools() {
+        return countTableRows("apptool");
+    }
+
+    private long countWorkflows() {
+        return countTableRows("workflow");
+    }
+
+    private long countNotebooks() {
+        return countTableRows("notebook");
+    }
+
+    private long countServices() {
+        return countTableRows("service");
+    }
+
+    private long countTableRows(String tableName) {
+        return testingPostgres.runSelectStatement("select count(*) from " + tableName, long.class);
+    }
+
+    private ApiException shouldThrowLambdaError(Runnable runnable) {
+        try {
+            runnable.run();
+            fail("should have thrown");
+            return null;
+        } catch (ApiException ex) {
+            assertEquals(LAMBDA_ERROR, ex.getCode());
+            return ex;
+        }
+    }
+
+    // the "multi-entry" repo has four .dockstore.yml entries
+    @Test
+    void testMultiEntryAllGood() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/master", USER_2_USERNAME);
+        assertEquals(2, countWorkflows());
+        assertEquals(2, countTools());
+    }
+
+    @Test
+    void testMultiEntryOneBroken() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        // test one broken tool
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/broken-tool", USER_2_USERNAME));
+        assertEquals(2, countWorkflows());
+        assertEquals(1, countTools());
+
+        // test one broken workflow
+        CommonTestUtilities.cleanStatePrivate2(SUPPORT, false, testingPostgres);
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/broken-workflow", USER_2_USERNAME));
+        assertEquals(1, countWorkflows());
+        assertEquals(2, countTools());
+    }
+
+    private void confirmNoEntries() {
+        assertEquals(0, countWorkflows());
+        assertEquals(0, countTools());
+        assertEquals(0, countNotebooks());
+        assertEquals(0, countServices());
+    }
+
+    @Test
+    void testMultiEntrySameName() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        // test tool-tool name collision
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/same-name-tool-tool", USER_2_USERNAME));
+        confirmNoEntries();
+
+        // test workflow-workflow name collision
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/same-name-workflow-workflow", USER_2_USERNAME));
+        confirmNoEntries();
+
+        // test tool-workflow name collision
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/same-name-tool-workflow", USER_2_USERNAME));
+        confirmNoEntries();
+
+        // test no names
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/no-names", USER_2_USERNAME));
+        confirmNoEntries();
+
+        // test service and unnamed workflows
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/service-and-unnamed-workflow", USER_2_USERNAME));
+        confirmNoEntries();
+
+        // test same name notebook-workflow collision
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/same-name-notebook-workflow", USER_2_USERNAME));
+        confirmNoEntries();
+
+        // test no name notebook-workflow collision
+        shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/no-name-notebook-workflow", USER_2_USERNAME));
+        confirmNoEntries();
+    }
+
+    /**
+     * Test that the push will fail if the .dockstore.yml contains a
+     * relative primary descriptor path, and the primary descriptor
+     * contains a relative secondary descriptor path.
+     */
+    @Test
+    void testMultiEntryRelativePrimaryDescriptorPath() {
+        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+        UsersApi usersApi = new UsersApi(webClient);
+        WorkflowsApi client = new WorkflowsApi(webClient);
+
+        ApiException ex = shouldThrowLambdaError(() -> handleGitHubRelease(client, INSTALLATION_ID, DockstoreTestingRepos.MULTI_ENTRY, "refs/heads/relative-primary-descriptor-path", USER_2_USERNAME));
+        assertTrue(ex.getMessage().toLowerCase().contains("could not be processed"));
+        assertEquals(0, countWorkflows());
+        assertEquals(2, countTools());
+        List<LambdaEvent> failedLambdaEvents = usersApi.getUserGitHubEvents(0, 10).stream()
+                .filter(event -> !event.isSuccess())
+                .toList();
+        assertEquals(2, failedLambdaEvents.size(), "There should be two failed events");
+        failedLambdaEvents.forEach(event -> assertTrue(event.getMessage().toLowerCase().contains("absolute"), "Should contain the word 'absolute'"));
     }
 }
